test: reduce redundant gap coverage suites and dedupe ProdStyle runs
- Exclude Category=ProdStyle on second pass in test-framework-prod-first
  and test-prime-time-full to avoid running ProdStyle tests twice
- Merge duplicate gap/main test pairs for JWT, HTTP barrier middleware,
  sanitizing provider factory, domain coverage, API key resolver,
  default barrier pipeline, and no-op audit sink
- Consolidate HttpBarrierContextMiddleware into a single test class
- Restore domain 100% line coverage for ImmutableCoreViolationException ctor
