+++ src/Nexo.Tests.BackgroundAgents/Trust/SanitizingProviderFactoryTests.cs
@@ -8,7 +8,7 @@ using Xunit;
 
 namespace Nexo.Tests.BackgroundAgents.Trust;
 
-public class SanitizingProviderFactoryTests
+public sealed class SanitizingProviderFactoryTests
 {
     [Fact]
     public async Task ExecuteLLMAsync_WhenProxyAllows_DelegatesToInner()
@@ -33,7 +33,7 @@ public class SanitizingProviderFactoryTests
     public async Task ExecuteLLMAsync_WhenProxyBlocks_Throws()
     {
         var innerMock = new Mock<IProviderFactory>();
-        var filter = new Nexo.BackgroundAgents.WebSearch.SensitiveContentFilter();
+        var filter = new SensitiveContentFilter();
         var proxy = new CloudSanitizationProxy(filter);
         var factory = new SanitizingProviderFactory(
             innerMock.Object,
@@ -45,6 +45,147 @@ public class SanitizingProviderFactoryTests
         await act.Should().ThrowAsync<InvalidOperationException>()
             .WithMessage("*blocked*");
         innerMock.Verify(x => x.ExecuteLLMAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<object>(), It.IsAny<CancellationToken>()), Times.Never);
+    }
+
+    [Fact]
+    public async Task ExecuteLLMAsync_BlocksWhenSanitizerDenies()
+    {
+        var inner = new Mock<IProviderFactory>();
+        var proxy = new Mock<ICloudSanitizationProxy>();
+        proxy.Setup(p => p.SanitizeForCloud(It.IsAny<OutgoingContext>(), It.IsAny<CancellationToken>()))
+            .Returns(SanitizationResult.Blocked("blocked"));
+
+        var factory = new SanitizingProviderFactory(inner.Object, proxy.Object, NullLogger<SanitizingProviderFactory>.Instance);
+
+        var act = () => factory.ExecuteLLMAsync("openai", "sys", "user", new { }, CancellationToken.None);
+        await act.Should().ThrowAsync<InvalidOperationException>().WithMessage("*blocked*");
+        inner.Verify(i => i.ExecuteLLMAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<object>(), It.IsAny<CancellationToken>()), Times.Never);
+    }
+
+    [Fact]
+    public async Task ExecuteLLMAsync_DelegatesSanitizedPrompts()
+    {
+        var inner = new Mock<IProviderFactory>();
+        inner.Setup(i => i.ExecuteLLMAsync("openai", "clean-sys", "clean-user", It.IsAny<object>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync("ok");
+
+        var proxy = new Mock<ICloudSanitizationProxy>();
+        proxy.Setup(p => p.SanitizeForCloud(It.IsAny<OutgoingContext>(), It.IsAny<CancellationToken>()))
+            .Returns(SanitizationResult.AllowedWith(new OutgoingContext
+            {
+                SystemPrompt = "clean-sys",
+                UserPrompt = "clean-user",
+                Provider = "openai",
+            }));
+
+        var factory = new SanitizingProviderFactory(inner.Object, proxy.Object, NullLogger<SanitizingProviderFactory>.Instance);
+        var result = await factory.ExecuteLLMAsync("openai", "dirty-sys", "dirty-user", new { }, CancellationToken.None);
+
+        result.Should().Be("ok");
+    }
+
+    [Fact]
+    public async Task ExecuteVisionAsync_BlocksWhenSanitizerDenies()
+    {
+        var inner = new Mock<IProviderFactory>();
+        var proxy = new Mock<ICloudSanitizationProxy>();
+        proxy.Setup(p => p.SanitizeForCloud(It.IsAny<OutgoingContext>(), It.IsAny<CancellationToken>()))
+            .Returns(SanitizationResult.Blocked("vision blocked"));
+
+        var factory = new SanitizingProviderFactory(inner.Object, proxy.Object, NullLogger<SanitizingProviderFactory>.Instance);
+        var act = () => factory.ExecuteVisionAsync("openai", "sys", "user", new byte[] { 1 }, new { }, CancellationToken.None);
+        await act.Should().ThrowAsync<InvalidOperationException>().WithMessage("*vision blocked*");
+    }
+
+    [Fact]
+    public async Task ExecuteVisionAsync_DelegatesWhenAllowed()
+    {
+        var inner = new Mock<IProviderFactory>();
+        inner.Setup(i => i.ExecuteVisionAsync("openai", "sys", "user", It.IsAny<byte[]>(), It.IsAny<object>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync("vision-ok");
+
+        var proxy = new Mock<ICloudSanitizationProxy>();
+        proxy.Setup(p => p.SanitizeForCloud(It.IsAny<OutgoingContext>(), It.IsAny<CancellationToken>()))
+            .Returns(SanitizationResult.AllowedWith(new OutgoingContext
+            {
+                SystemPrompt = "sys",
+                UserPrompt = "user",
+                Provider = "openai",
+            }));
+
+        var factory = new SanitizingProviderFactory(inner.Object, proxy.Object, NullLogger<SanitizingProviderFactory>.Instance);
+        var result = await factory.ExecuteVisionAsync("openai", "sys", "user", new byte[] { 1 }, new { }, CancellationToken.None);
+
+        result.Should().Be("vision-ok");
+    }
+
+    [Fact]
+    public async Task ExecuteVisionMultiFrameAsync_DelegatesWhenAllowed()
+    {
+        var inner = new Mock<IProviderFactory>();
+        inner.Setup(i => i.ExecuteVisionMultiFrameAsync(
+                "openai",
+                "sys",
+                "user",
+                It.IsAny<IReadOnlyList<byte[]>>(),
+                It.IsAny<object>(),
+                It.IsAny<CancellationToken>()))
+            .ReturnsAsync("frames");
+
+        var proxy = new Mock<ICloudSanitizationProxy>();
+        proxy.Setup(p => p.SanitizeForCloud(It.IsAny<OutgoingContext>(), It.IsAny<CancellationToken>()))
+            .Returns(SanitizationResult.AllowedWith(new OutgoingContext
+            {
+                SystemPrompt = "sys",
+                UserPrompt = "user",
+                Provider = "openai",
+            }));
+
+        var factory = new SanitizingProviderFactory(inner.Object, proxy.Object, NullLogger<SanitizingProviderFactory>.Instance);
+        var result = await factory.ExecuteVisionMultiFrameAsync(
+            "openai",
+            "sys",
+            "user",
+            new[] { new byte[] { 1 }, new byte[] { 2 } },
+            new { },
+            CancellationToken.None);
+
+        result.Should().Be("frames");
+    }
+
+    [Fact]
+    public async Task ExecuteVideoAsync_DelegatesWhenAllowed()
+    {
+        var inner = new Mock<IProviderFactory>();
+        inner.Setup(i => i.ExecuteVideoAsync("clean-sys", "clean-user", It.IsAny<IReadOnlyList<byte[]>>(), It.IsAny<object>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync("video-ok");
+
+        var proxy = new Mock<ICloudSanitizationProxy>();
+        proxy.Setup(p => p.SanitizeForCloud(It.IsAny<OutgoingContext>(), It.IsAny<CancellationToken>()))
+            .Returns(SanitizationResult.AllowedWith(new OutgoingContext
+            {
+                SystemPrompt = "clean-sys",
+                UserPrompt = "clean-user",
+                Provider = "video",
+            }));
+
+        var factory = new SanitizingProviderFactory(inner.Object, proxy.Object, NullLogger<SanitizingProviderFactory>.Instance);
+        var result = await factory.ExecuteVideoAsync("dirty-sys", "dirty-user", new[] { new byte[] { 1 } }, new { }, CancellationToken.None);
+
+        result.Should().Be("video-ok");
+    }
+
+    [Fact]
+    public async Task ExecuteVideoAsync_BlocksWhenSanitizerDenies()
+    {
+        var inner = new Mock<IProviderFactory>();
+        var proxy = new Mock<ICloudSanitizationProxy>();
+        proxy.Setup(p => p.SanitizeForCloud(It.IsAny<OutgoingContext>(), It.IsAny<CancellationToken>()))
+            .Returns(SanitizationResult.Blocked("video blocked"));
+
+        var factory = new SanitizingProviderFactory(inner.Object, proxy.Object, NullLogger<SanitizingProviderFactory>.Instance);
+        var act = () => factory.ExecuteVideoAsync("sys", "user", new[] { new byte[] { 1 } }, new { }, CancellationToken.None);
+        await act.Should().ThrowAsync<InvalidOperationException>().WithMessage("*video blocked*");
     }
 
     [Fact]
