+++ src/Nexo.Tests.Infrastructure/Barriers/Identity/ApiKeyBarrierResolverTests.cs
@@ -11,6 +11,23 @@ namespace Nexo.Tests.Infrastructure.Barriers.Identity;
 public sealed class ApiKeyBarrierResolverTests
 {
     [Fact]
+    public void Constructor_throws_for_null_dependencies()
+    {
+        var options = new ApiKeyResolverOptions();
+        var hierarchy = CreateHierarchy();
+        var logger = new TestLogger<ApiKeyBarrierResolver>();
+
+        var act = () => new ApiKeyBarrierResolver(null!, hierarchy, logger);
+        act.Should().Throw<ArgumentNullException>().WithParameterName("options");
+
+        act = () => new ApiKeyBarrierResolver(options, null!, logger);
+        act.Should().Throw<ArgumentNullException>().WithParameterName("hierarchy");
+
+        act = () => new ApiKeyBarrierResolver(options, hierarchy, null!);
+        act.Should().Throw<ArgumentNullException>().WithParameterName("logger");
+    }
+
+    [Fact]
     public async Task TryResolveAsync_ValidHashedKeyMapping_ReturnsLevel()
     {
         var key = "test-key-123";
@@ -103,6 +120,170 @@ public sealed class ApiKeyBarrierResolverTests
         var result = await sut.TryResolveAsync(CreateContext(apiKey: null));
 
         result.Should().BeNull();
+    }
+
+    [Fact]
+    public async Task TryResolveAsync_WhitespaceApiKey_ReturnsNull()
+    {
+        var sut = CreateResolver(new ApiKeyResolverOptions
+        {
+            KeyMapping = new Dictionary<string, string>
+            {
+                [ComputeSha256Hex("valid")] = "public",
+            },
+        });
+
+        var result = await sut.TryResolveAsync(CreateContext("   "));
+
+        result.Should().BeNull();
+    }
+
+    [Fact]
+    public async Task TryResolveAsync_UsesDefaultHeaderNameWhenContextApiKeyMissing()
+    {
+        var key = "default-header-key";
+        var hash = ComputeSha256Hex(key);
+        var sut = CreateResolver(new ApiKeyResolverOptions
+        {
+            KeyMapping = new Dictionary<string, string> { [hash] = "internal" },
+        });
+
+        var result = await sut.TryResolveAsync(new BarrierResolutionContext(
+            CorrelationId: "corr",
+            ExplicitLevel: null,
+            Headers: new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
+            {
+                ["x-nexo-api-key"] = key,
+            },
+            CertSubjects: Array.Empty<string>(),
+            CertSans: Array.Empty<string>(),
+            RawJwt: null,
+            JwtClaims: new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase),
+            ApiKey: null));
+
+        result.Should().NotBeNull();
+        result!.ResolvedLevel.Should().Be("internal");
+    }
+
+    [Fact]
+    public async Task TryResolveAsync_NullContext_ReturnsNull()
+    {
+        var sut = CreateResolver(new ApiKeyResolverOptions());
+
+        var result = await sut.TryResolveAsync(null!);
+
+        result.Should().BeNull();
+    }
+
+    [Fact]
+    public async Task TryResolveAsync_ReadsApiKeyFromConfiguredHeaderName()
+    {
+        var key = "header-key";
+        var hash = ComputeSha256Hex(key);
+        var sut = CreateResolver(new ApiKeyResolverOptions
+        {
+            HeaderName = "x-custom-api-key",
+            KeyMapping = new Dictionary<string, string> { [hash] = "internal" },
+        });
+
+        var result = await sut.TryResolveAsync(new BarrierResolutionContext(
+            CorrelationId: "corr",
+            ExplicitLevel: null,
+            Headers: new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
+            {
+                ["x-custom-api-key"] = key,
+            },
+            CertSubjects: Array.Empty<string>(),
+            CertSans: Array.Empty<string>(),
+            RawJwt: null,
+            JwtClaims: new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase),
+            ApiKey: null));
+
+        result.Should().NotBeNull();
+        result!.ResolvedLevel.Should().Be("internal");
+    }
+
+    [Fact]
+    public async Task TryResolveAsync_UnknownMappedLevel_LogsWarningAndReturnsNull()
+    {
+        var key = "abc";
+        var hash = ComputeSha256Hex(key);
+        var logger = new TestLogger<ApiKeyBarrierResolver>();
+        var sut = new ApiKeyBarrierResolver(
+            new ApiKeyResolverOptions { KeyMapping = new Dictionary<string, string> { [hash] = "missing-level" } },
+            CreateHierarchy(),
+            logger);
+
+        var result = await sut.TryResolveAsync(CreateContext(key));
+
+        result.Should().BeNull();
+        logger.Entries.Should().Contain(entry =>
+            entry.Level == Microsoft.Extensions.Logging.LogLevel.Warning &&
+            entry.Message.Contains("unknown barrier level", StringComparison.OrdinalIgnoreCase));
+    }
+
+    [Fact]
+    public async Task TryResolveAsync_ShortKey_UsesFullPrefixInDetail()
+    {
+        var key = "ab";
+        var hash = ComputeSha256Hex(key);
+        var sut = CreateResolver(new ApiKeyResolverOptions
+        {
+            KeyMapping = new Dictionary<string, string> { [hash] = "public" },
+        });
+
+        var result = await sut.TryResolveAsync(CreateContext(key));
+
+        result.Should().NotBeNull();
+        result!.Detail.Should().Contain("ab****");
+    }
+
+    [Fact]
+    public async Task TryResolveAsync_SkipsBlankConfiguredMappingHash()
+    {
+        var key = "live-key";
+        var hash = ComputeSha256Hex(key);
+        var sut = CreateResolver(new ApiKeyResolverOptions
+        {
+            KeyMapping = new Dictionary<string, string>
+            {
+                [""] = "internal",
+                [hash] = "public",
+            },
+        });
+
+        var result = await sut.TryResolveAsync(CreateContext(key));
+
+        result.Should().NotBeNull();
+        result!.ResolvedLevel.Should().Be("public");
+    }
+
+    [Fact]
+    public async Task TryResolveAsync_PrefersContextApiKeyOverHeader()
+    {
+        var contextKey = "context-key";
+        var headerKey = "header-key";
+        var hash = ComputeSha256Hex(contextKey);
+        var sut = CreateResolver(new ApiKeyResolverOptions
+        {
+            KeyMapping = new Dictionary<string, string> { [hash] = "internal" },
+        });
+
+        var result = await sut.TryResolveAsync(new BarrierResolutionContext(
+            CorrelationId: "corr",
+            ExplicitLevel: null,
+            Headers: new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
+            {
+                ["x-nexo-api-key"] = headerKey,
+            },
+            CertSubjects: Array.Empty<string>(),
+            CertSans: Array.Empty<string>(),
+            RawJwt: null,
+            JwtClaims: new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase),
+            ApiKey: contextKey));
+
+        result.Should().NotBeNull();
+        result!.ResolvedLevel.Should().Be("internal");
     }
 
     private static ApiKeyBarrierResolver CreateResolver(ApiKeyResolverOptions options)
