+++ src/Nexo.Tests.Infrastructure/Barriers/Identity/DefaultBarrierIdentityResolverPipelineTests.cs
@@ -10,6 +10,32 @@ namespace Nexo.Tests.Infrastructure.Barriers.Identity;
 public sealed class DefaultBarrierIdentityResolverPipelineTests
 {
     [Fact]
+    public void Constructor_throws_for_null_logger()
+    {
+        var act = () => new DefaultBarrierIdentityResolverPipeline([], null!);
+
+        act.Should().Throw<ArgumentNullException>().WithParameterName("logger");
+    }
+
+    [Fact]
+    public void Constructor_treats_null_resolver_enumerable_as_empty()
+    {
+        var sut = new DefaultBarrierIdentityResolverPipeline(null!, new TestLogger<DefaultBarrierIdentityResolverPipeline>());
+
+        sut.Should().NotBeNull();
+    }
+
+    [Fact]
+    public async Task ResolveAsync_throws_for_null_context()
+    {
+        var sut = new DefaultBarrierIdentityResolverPipeline([], new TestLogger<DefaultBarrierIdentityResolverPipeline>());
+
+        var act = async () => await sut.ResolveAsync(null!);
+
+        await act.Should().ThrowAsync<ArgumentNullException>().WithParameterName("context");
+    }
+
+    [Fact]
     public async Task ResolveAsync_ExplicitLevel_ReturnsImmediately_WithoutCallingResolvers()
     {
         var resolver = new Mock<IBarrierIdentityResolver>(MockBehavior.Strict);
@@ -120,6 +146,46 @@ public sealed class DefaultBarrierIdentityResolverPipelineTests
         result.Should().BeNull();
     }
 
+    [Fact]
+    public async Task ResolveAsync_WhitespaceExplicitLevel_FallsThroughToResolvers()
+    {
+        var resolver = new StubResolver("stub", new BarrierResolutionResult("internal", "stub", BarrierAuthoritySource.ApiKey));
+        var sut = new DefaultBarrierIdentityResolverPipeline([resolver], new TestLogger<DefaultBarrierIdentityResolverPipeline>());
+
+        var result = await sut.ResolveAsync(CreateContext(explicitLevel: "   "));
+
+        result.Should().NotBeNull();
+        result!.ResolverName.Should().Be("stub");
+    }
+
+    [Fact]
+    public async Task ResolveAsync_ExplicitLevelWithoutBarrierHeader_UsesCliAuthority()
+    {
+        var sut = new DefaultBarrierIdentityResolverPipeline([], new TestLogger<DefaultBarrierIdentityResolverPipeline>());
+
+        var result = await sut.ResolveAsync(CreateContext(explicitLevel: "internal"));
+
+        result.Should().NotBeNull();
+        result!.AuthoritySource.Should().Be(BarrierAuthoritySource.Cli);
+        result.Detail.Should().Contain("Explicit barrier provided");
+    }
+
+    [Fact]
+    public async Task ResolveAsync_AllResolversThrow_ReturnsNullAndLogsWarnings()
+    {
+        var logger = new TestLogger<DefaultBarrierIdentityResolverPipeline>();
+        var sut = new DefaultBarrierIdentityResolverPipeline(
+            [new ThrowingResolver("first"), new ThrowingResolver("second")],
+            logger);
+
+        var result = await sut.ResolveAsync(CreateContext());
+
+        result.Should().BeNull();
+        logger.Entries.Count(entry =>
+            entry.Level == Microsoft.Extensions.Logging.LogLevel.Warning &&
+            entry.Message.Contains("resolver failed", StringComparison.OrdinalIgnoreCase)).Should().Be(2);
+    }
+
     private static BarrierResolutionContext CreateContext(string? explicitLevel = null)
         => new(
             CorrelationId: "corr-1",
@@ -157,6 +223,6 @@ public sealed class DefaultBarrierIdentityResolverPipelineTests
         public ValueTask<BarrierResolutionResult?> TryResolveAsync(
             BarrierResolutionContext context,
             CancellationToken cancellationToken = default)
-            => throw new InvalidOperationException("boom");
+            => throw new InvalidOperationException("resolver failed");
     }
 }
